+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   \#10: frodo-lib is now a hybrid npm package supporting both ES modules and CommonJS modules.
+
 ## [0.11.1-3] - 2022-08-18
 
 ## [0.11.1-2] - 2022-08-18